+++ server/Hash.h
@@ -1,6 +1,6 @@
 // Copyright (C)2004 Landmark Graphics Corporation
 // Copyright (C)2005 Sun Microsystems, Inc.
-// Copyright (C)2011, 2014 D. R. Commander
+// Copyright (C)2011, 2014, 2019 D. R. Commander
 //
 // This library is free software and may be redistributed and/or modified under
 // the terms of the wxWindows Library License, Version 3.1 or (at your option)
@@ -134,7 +134,7 @@ namespace vglserver
 				if(entry->next) entry->next->prev = entry->prev;
 				if(entry == start) start = entry->next;
 				if(entry == end) end = entry->prev;
-				if(entry->value) detach(entry);
+				detach(entry);
 				memset(entry, 0, sizeof(HashEntry));
 				delete entry;
 				count--;