+++ server/VirtualPixmap.cpp
@@ -1,4 +1,4 @@
-// Copyright (C)2011, 2013-2014, 2017 D. R. Commander
+// Copyright (C)2011, 2013-2014, 2017, 2019 D. R. Commander
 //
 // This library is free software and may be redistributed and/or modified under
 // the terms of the wxWindows Library License, Version 3.1 or (at your option)
@@ -20,12 +20,12 @@ using namespace vglcommon;
 using namespace vglserver;
 
 
-VirtualPixmap::VirtualPixmap(Display *dpy_, XVisualInfo *vis, Pixmap pm) :
+VirtualPixmap::VirtualPixmap(Display *dpy_, Visual *visual, Pixmap pm) :
 	VirtualDrawable(dpy_, pm)
 {
 	CriticalSection::SafeLock l(mutex);
 	profPMBlit.setName("PMap Blit ");
-	_newcheck(frame = new FBXFrame(dpy_, pm, vis->visual, true));
+	_newcheck(frame = new FBXFrame(dpy_, pm, visual, true));
 }